% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hu_codebook.R
\name{get_hu_codebook}
\alias{get_hu_codebook}
\title{Get the housing unit variable codebook}
\usage{
get_hu_codebook()
}
\value{
A data frame with columns: \code{variable}, \code{label},
\code{unit}, \code{nominal}, \code{notes}.
}
\description{
Returns a data frame describing each variable available from
\code{\link{read_hu_from_s3}}.
}
\examples{
get_hu_codebook()

}
\seealso{
\code{\link{read_hu_from_s3}}
}
